[S12P21E206-266] Fix: 유저연령대리스트 수정

+++ back/user/src/main/java/com/ssafy/user/service/impl/UserServiceImpl.java
@@ -226,7 +226,7 @@ public class UserServiceImpl implements UserService {
         // 매퍼에서 결과는 ageGroup(0~5)와 count로 구성된 Map 리스트로 반환됨
         List<Map<String, Object>> results = userMapper.selectAgeGroupCounts(dtos);
         for (Map<String, Object> row : results) {
-            int group = (int) row.get("ageGroup");
+            int group = ((Long) row.get("ageGroup")).intValue();
             // count를 Long에서 int로 변환
 //            int count = ((Long) row.get("count")).intValue();
             Long count = (Long) row.get("count");
